support specifying network name with `docker run --net`
Currently the only way to use a network created with `docker network create` is
the --publish-service option. If the extra flexibility of services is not
required, then that option has the disadvantage of making the user repeat the
container name. Another problem is that the service created persists after the
container is deleted.

This change enables usage like `docker network create blue; docker run --net
blue mycontainer`.

Signed-off-by: Rich Lane <rlane@bigswitch.com>